test: Move 요구사항_3 test of CarTest class to CarNameTest class
