+++ src/test/java/racingcar/domain/CarTest.java
@@ -1,10 +1,8 @@
-package racingcar;
+package racingcar.domain;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import racingcar.domain.Car;
-import racingcar.domain.CarName;
 
 import static org.assertj.core.api.Assertions.*;
